Moved setup to Doctor
